+++ crates/rpc-types-trace/src/geth/mod.rs
@@ -23,6 +23,11 @@ pub mod mux;
 pub mod noop;
 pub mod pre_state;
 
+/// Error when the inner tracer from [GethTrace] is mismatching to the target tracer.
+#[derive(Debug, thiserror::Error)]
+#[error("unexpected tracer")]
+pub struct UnexpectedTracerError(pub GethTrace);
+
 /// Result type for geth style transaction trace
 pub type TraceResult = crate::common::TraceResult<GethTrace, String>;
 
@@ -121,6 +126,64 @@ pub enum GethTrace {
     MuxTracer(MuxFrame),
     /// Any other trace response, such as custom javascript response objects
     JS(serde_json::Value),
+}
+
+impl GethTrace {
+    /// Try to convert the inner tracer to [DefaultFrame]
+    pub fn try_into_default_frame(self) -> Result<DefaultFrame, UnexpectedTracerError> {
+        match self {
+            Self::Default(inner) => Ok(inner),
+            _ => Err(UnexpectedTracerError(self)),
+        }
+    }
+
+    /// Try to convert the inner tracer to [CallFrame]
+    pub fn try_into_call_frame(self) -> Result<CallFrame, UnexpectedTracerError> {
+        match self {
+            Self::CallTracer(inner) => Ok(inner),
+            _ => Err(UnexpectedTracerError(self)),
+        }
+    }
+
+    /// Try to convert the inner tracer to [FourByteFrame]
+    pub fn try_into_four_byte_frame(self) -> Result<FourByteFrame, UnexpectedTracerError> {
+        match self {
+            Self::FourByteTracer(inner) => Ok(inner),
+            _ => Err(UnexpectedTracerError(self)),
+        }
+    }
+
+    /// Try to convert the inner tracer to [PreStateFrame]
+    pub fn try_into_pre_state_frame(self) -> Result<PreStateFrame, UnexpectedTracerError> {
+        match self {
+            Self::PreStateTracer(inner) => Ok(inner),
+            _ => Err(UnexpectedTracerError(self)),
+        }
+    }
+
+    /// Try to convert the inner tracer to [NoopFrame]
+    pub fn try_into_noop_frame(self) -> Result<NoopFrame, UnexpectedTracerError> {
+        match self {
+            Self::NoopTracer(inner) => Ok(inner),
+            _ => Err(UnexpectedTracerError(self)),
+        }
+    }
+
+    /// Try to convert the inner tracer to [MuxFrame]
+    pub fn try_into_mux_frame(self) -> Result<MuxFrame, UnexpectedTracerError> {
+        match self {
+            Self::MuxTracer(inner) => Ok(inner),
+            _ => Err(UnexpectedTracerError(self)),
+        }
+    }
+
+    /// Try to convert the inner tracer to [serde_json::Value]
+    pub fn try_into_json_value(self) -> Result<serde_json::Value, UnexpectedTracerError> {
+        match self {
+            Self::JS(inner) => Ok(inner),
+            _ => Err(UnexpectedTracerError(self)),
+        }
+    }
 }
 
 impl Default for GethTrace {
@@ -658,5 +721,44 @@ mod tests {
         let de = serde_json::to_value(&result).unwrap();
         let val = serde_json::from_str::<serde_json::Value>(s).unwrap();
         similar_asserts::assert_eq!(val, de);
+    }
+
+    #[test]
+    fn test_geth_trace_into_tracer() {
+        let geth_trace = GethTrace::Default(DefaultFrame::default());
+        let inner = geth_trace.try_into_default_frame();
+        assert!(inner.is_ok());
+
+        let geth_trace = GethTrace::CallTracer(CallFrame::default());
+        let inner = geth_trace.try_into_call_frame();
+        assert!(inner.is_ok());
+
+        let geth_trace = GethTrace::FourByteTracer(FourByteFrame::default());
+        let inner = geth_trace.try_into_four_byte_frame();
+        assert!(inner.is_ok());
+
+        let geth_trace = GethTrace::PreStateTracer(PreStateFrame::Default(PreStateMode::default()));
+        let inner = geth_trace.try_into_pre_state_frame();
+        assert!(inner.is_ok());
+
+        let geth_trace = GethTrace::NoopTracer(NoopFrame::default());
+        let inner = geth_trace.try_into_noop_frame();
+        assert!(inner.is_ok());
+
+        let geth_trace = GethTrace::MuxTracer(MuxFrame::default());
+        let inner = geth_trace.try_into_mux_frame();
+        assert!(inner.is_ok());
+
+        let geth_trace = GethTrace::JS(serde_json::Value::Null);
+        let inner = geth_trace.try_into_json_value();
+        assert!(inner.is_ok());
+    }
+
+    #[test]
+    fn test_geth_trace_into_tracer_wrong_tracer() {
+        let geth_trace = GethTrace::Default(DefaultFrame::default());
+        let inner = geth_trace.try_into_call_frame();
+        assert!(inner.is_err());
+        assert!(matches!(inner, Err(UnexpectedTracerError(_))));
     }
 }